Search using single union  query (#2709)
https://eaflood.atlassian.net/browse/WATER-5292

We are migrating the internal search page (the root page of the internal service) from the legacy apps to our new `water-abstraction-system` app.

Each different type of search has been migrated individually so far.

This change is to combine them into a single, simpler query that will display all the results in a logical sequence, correctly paginated.

It also includes some changes to how the results are displayed.

This should be the last major change to implement the new searching facility. Once this has been reviewed, amended and merged, another change will be required to this code and the legacy code to add a feature flag to use the new search pages instead of the existing one.