Fix missed returns when licence ends on start date
https://eaflood.atlassian.net/browse/WATER-5435

The B&D team marked a licence as expired on 1 July 2025 in NALD. The licence submits quarterly returns so this expiry date lines up with the start date of the second quarterly period.

When the overnight import job processed this, they expected

- existing return logs for quarters 2, 3 and 4 to be voided
- a new return log for 2025-07-01 to 2025-07-01 (one day) to be created

The voiding happened, but no new return log was created. Their expectation is correct and after investigation we found the issue was an inconsistent check in `CreateReturnLogsService`.

```javascript
  // Good - equal to or greater than
  const startDateInPeriod = returnRequirement.returnVersion.startDate <= period.endDate

  // Good - equal to or greater than
  const endDateInPeriod = returnRequirement.returnVersion.endDate >= period.startDate

  // Bad - only greater than
  const licenceEndDateInPeriod = licenceEndDate > period.startDate
```

As the start date for the second quarter is 2025-07-01 and the licence 'end date' is '2025-07-01', `CreateReturnLogsService` determines the second quarter does not need processing.

This change fixes the inconsistency