First step in making testing clearer
The fix is literally adding the character `=` to the code. But we want to add a test to ensure this scenario is covered in the future.

When we came to do this we realised the way the fixtures are currently structured means its not clear on what basis the return logs are being created.

So, we're going to try and make the existing tests clearer, by splitting out and naming the individual fixture records. We'll then do some clean up and remove duplication, and use generated data where possible.

This first step takes the two requirement fixtures that were being created and moves them to named functions.