Remove redundant returnRequirementsAcrossReturnVersions()
This fixture appears to have been added to test the logic in `ProcessLicenceReturnLogsService` that matches return requirements to the return cycle currently being processed.

What it flags is that the only thing that really matters is whether the return requirement is summer, and the start and end date of the return version inside the return requirement.

In fact, what was previously being returned does not accurately represent the return versions for a licence, because the dates not flow properly together.

For now, we've replicated what was needed by the tests to allow us to remove the fixture function, and a lot of duplicated code.