Remove deduplicate licence feature
https://eaflood.atlassian.net/browse/WATER-4429

In [Add a remove duplicate licence feature](https://github.com/DEFRA/water-abstraction-system/pull/883) we added a 'secret' feature that allowed us to deal with a data issue that was raised once or twice a month.

The service would have problems with a licence because there would be two records with the same reference, just one would feature additional whitespace before or after the reference.

We believe when a change is made to a licence in NALD it requires users to enter the licence reference again. This creates a new record in NALD but our import service recognises it is for the same licence because the references match.

However, NALD is known for poor validation and doesn't trim data entered. So, often a user will accidentally add an extra space or newline character when entering the licence reference again. On import, computers being the way they are, will see this as a new licence so will create a new licence record with the same 'bad' reference.

This causes an error for anything that attempts to retrieve the licence by reference in the service. The most visible casualty is the search page.

At the time we said we _should_ be fixing [water-abstraction-import](https://github.com/DEFRA/water-abstraction-import). But we were still new to the service so were not ready to dive in.

Fast forward a couple of years and we have essentially re-written the import process. We have also added functionality to ensure that data deleted in NALD is deleted in WRLS. This includes licences. In the case of these whitespace records, the issue is spotted and the reference corrected in NALD.

This means the import now matches the two licence records (NALD and WRLS) correctly. When the import is complete it sees the one with whitespace has not been updated and does not exist in NALD anymore so deletes it.

So we don't need this endpoint anymore or the feature behind it.