Remove deduplicate licence feature (#2817)
https://eaflood.atlassian.net/browse/WATER-4429

In [Add a remove duplicate licence feature](https://github.com/DEFRA/water-abstraction-system/pull/883), we added a 'secret' feature that addressed a data issue raised once or twice a month.

The service would have problems with a licence because there would be two records with the same reference, one would feature additional whitespace before or after the reference.

When a change is made to a licence in NALD, users must re-enter the licence reference. This creates a new record in NALD, but our import service recognises it is for the same licence because the references match.

However, NALD is known for poor validation and doesn't trim data entered. So, a user will often accidentally add an extra space or newline when entering the licence reference again. On import, computers, being the way they are, will see this as a new licence, so they will create a new licence record with the same 'bad' reference.

This causes an error for any attempt to retrieve the licence by reference in the service. The most visible casualty is the search page.

At the time, we said we _should_ be fixing [water-abstraction-import](https://github.com/DEFRA/water-abstraction-import). But we were still new to the service, so we were not ready to dive in.

Fast forward a couple of years, and we have essentially rewritten the import process. We have also added functionality to ensure that data deleted in NALD is deleted in WRLS. This includes licences. For these whitespace records, the issue is identified and the reference corrected in NALD.

This means the import now correctly matches the two licence records (NALD and WRLS). When the import is complete, it sees that the one with whitespace has not been updated and no longer exists in NALD, so it deletes it.

So we don't need this endpoint or the feature it supports anymore.