Enable new system search (#2780)
https://eaflood.atlassian.net/browse/WATER-5292

We are migrating the internal search page (the root page of the internal service) from the legacy apps to our new `water-abstraction-system` app.

The new search page has been created and can be enabled using a new feature flag `USE_SYSTEM_SEARCH` in the `water-abstraction-ui` application.

This flag tells the application to use the new search page instead of the old one. The page is served from the system application at `/system/search`.
The `water-abstraction-ui` application redirects any requests to its root to the search page - when the feature flag is set that will be the new search page.

To avoid multiple redirects, any links in that point to the old  `/licences` path have been updated to point to the correct root path `/` instead.

This change is to perform a similar update in  `water-abstraction-system`, ensuring that links to  `/licences` point to the correct root path `/` instead.

It also adds an `id` to the search button on the new search page so that it can be easily located by our acceptance tests.